feat: EC0009 implement global exception handling for resource APIs with @ControllerAdvice

+++ src/main/java/com/oclock/event_backend/service/RefreshTokenServiceImpl.java
@@ -3,6 +3,8 @@ package com.oclock.event_backend.service;
 import com.oclock.event_backend.domain.RefreshToken;
 import com.oclock.event_backend.domain.User;
 import com.oclock.event_backend.dto.AuthResponse;
+import com.oclock.event_backend.exception.EmptyInputException;
+import com.oclock.event_backend.exception.FunctionalException;
 import com.oclock.event_backend.repository.RefreshTokenRepository;
 import com.oclock.event_backend.util.JwtUtil;
 import lombok.AllArgsConstructor;
@@ -34,10 +36,14 @@ public class RefreshTokenServiceImpl implements RefreshTokenService {
 
     @Override
     public AuthResponse refreshAccessToken(String refreshToken) {
+        if (refreshToken == null || refreshToken.isEmpty()) {
+            throw new EmptyInputException("Refresh token is required.");
+        }
+
         Optional<RefreshToken> optionalToken = findByToken(refreshToken);
 
-        if (optionalToken.isEmpty() || optionalToken.get().getExpiryDate().isBefore(Instant.now())) {
-            throw new IllegalArgumentException("Invalid or expired refresh token");
+        if (optionalToken.get().getExpiryDate().isBefore(Instant.now())) {
+            throw new FunctionalException("Expired refresh token.");
         }
 
         String newAccessToken = jwtUtil.generateAccessToken(optionalToken.get().getUser());
@@ -55,10 +61,14 @@ public class RefreshTokenServiceImpl implements RefreshTokenService {
 
     @Override
     public void deleteByToken(String refreshToken) {
+        if (refreshToken == null || refreshToken.isEmpty()) {
+            throw new EmptyInputException("Refresh token is required.");
+        }
+
         Optional<RefreshToken> optionalToken = refreshTokenRepository.findByToken(refreshToken);
 
         if (optionalToken.isEmpty()) {
-            throw new IllegalArgumentException("Invalid or missing refresh token");
+            throw new FunctionalException("Invalid refresh token.");
         }
 
         refreshTokenRepository.deleteByToken(refreshToken);
